# resources/openapi/linq.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "linq",
    "version": "0.1.0",
    "summary": "Self-hosted URL shortener and dynamic-link service.",
    "description": "Three things share one port: the REST API under `/api`, the Client UI's\nstatic export under `/home`, and the redirect on every other path.\n\n**Authentication.** Every `/api/v1/*` route needs an API key, sent as\n`Authorization: Bearer <key>` or `X-Api-Key: <key>`. A key is the principal:\nnobody logs in, and there are no user accounts. A key whose `expires_at` has\npassed fails with 401.\n\n**Roles** are ordered `viewer < author < manager < admin`. Each operation\nbelow states the minimum it needs.\n\n**Errors** all share one envelope, documented on the `Error` schema.\n\n**Lists** all share one envelope — `{ data, total, limit, offset }` — where\n`total` counts every row matching the filters, not just the page.\n\n---\n\nWritten by hand from `apps/server/src/http/`, which stays the source of\ntruth; the zod schemas in `packages/shared/src/` are what the server\nactually validates against, and the constraints here mirror them.\n"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://{host}",
      "description": "A deployed instance",
      "variables": {
        "host": {
          "default": "links.example.com"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "Identity"
        ],
        "operationId": "getMe",
        "summary": "The calling key's identity",
        "description": "The calling key. There is no user behind it: a key carries its own name and role, and this is what the Client UI gates its menus on.\n",
        "responses": {
          "200": {
            "description": "The calling key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKey"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/domains": {
      "get": {
        "tags": [
          "Domains"
        ],
        "operationId": "listDomains",
        "summary": "List domains",
        "description": "Ordered by host. Every role may read.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of domains.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Page"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Domain"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Domains"
        ],
        "operationId": "createDomain",
        "summary": "Create a domain",
        "description": "Requires: **admin**.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created domain.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Domain"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "The host already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/links": {
      "get": {
        "tags": [
          "Links"
        ],
        "operationId": "listLinks",
        "summary": "List links",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "name": "search",
            "in": "query",
            "description": "Substring match, case-insensitive, over slug, name and destination.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Comma-separated. A link matches when it carries **any** of them.",
            "schema": {
              "type": "string",
              "examples": [
                "launch,press"
              ]
            }
          },
          {
            "name": "domain_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "archived",
                "all"
              ],
              "default": "active"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "visits"
              ],
              "default": "created_at"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "expiry",
            "in": "query",
            "description": "Filters on expires_at against the current time. `live` includes links that never expire.",
            "schema": {
              "type": "string",
              "enum": [
                "any",
                "live",
                "expired"
              ],
              "default": "any"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of links.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Page"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Link"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Links"
        ],
        "operationId": "createLink",
        "summary": "Create a link",
        "description": "Requires: **editor**.\n\nOmit `slug` for a generated one. A custom slug that is taken on that\ndomain — including by an archived link — is a 409.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created link.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Link"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "The domain does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The slug is taken, the domain is archived, or no free slug could be allocated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/qr-codes": {
      "get": {
        "tags": [
          "QR codes"
        ],
        "operationId": "listQrCodes",
        "summary": "List QR codes",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "name": "link_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Substring match, case-insensitive, over the QR code's own name and its link's name and slug.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of QR codes.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Page"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/QrCode"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "QR codes"
        ],
        "operationId": "createQrCode",
        "summary": "Create a QR code",
        "description": "Requires: **editor**. A QR code is always attached to a link — `link_id` is required — and the link must be active: an archived link's short URL does not resolve, so a code for one would be born dead.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrCodeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created QR code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QrCode"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "The link does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The link is archived.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/visits": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "listVisits",
        "summary": "The raw visit log",
        "description": "Newest first, narrowed by `link_id`, `domain_id` or `orphan`. One route\nfor every scope, because the only thing that changes between them is a\npredicate.\n\nUnlike the reports above, this reads the detail table: a visit row is\nthe only place a user agent, a query string or an exact instant\nsurvives. Its window is therefore in **instants**, not whole days.\n\nAn id that matches nothing reports nothing rather than 404ing — this is\nthe log with a filter on it, not a view of one link.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `occurred_at`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Inclusive upper bound on `occurred_at`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "bot",
            "in": "query",
            "description": "`any` keeps bots and humans together.",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "any"
              ],
              "default": "any"
            }
          },
          {
            "name": "link_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "domain_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/Orphan"
          },
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Platform"
            }
          },
          {
            "name": "os",
            "in": "query",
            "description": "Open vocabulary (unlike platform), lowercased at the boundary.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "browser",
            "in": "query",
            "description": "Open vocabulary (unlike platform), lowercased at the boundary.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of visits.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Page"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Visit"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/tags": {
      "get": {
        "tags": [
          "Links"
        ],
        "operationId": "listTags",
        "summary": "Tags in use, with counts",
        "description": "Derived from active links; there is no tag table to keep in step.\nBusiest first, then alphabetical.\n",
        "responses": {
          "200": {
            "description": "Every tag carried by at least one active link.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "tag",
                      "count"
                    ],
                    "properties": {
                      "tag": {
                        "type": "string"
                      },
                      "count": {
                        "type": "integer",
                        "minimum": 1
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/keys": {
      "get": {
        "operationId": "listKeys",
        "tags": [
          "Keys"
        ],
        "summary": "List keys",
        "description": "Every role may list keys, but only an admin sees more than `{ id, name, role }`.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of keys.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Page"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "anyOf": [
                              {
                                "$ref": "#/components/schemas/ApiKey"
                              },
                              {
                                "$ref": "#/components/schemas/ApiKeySummary"
                              }
                            ]
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createKey",
        "tags": [
          "Keys"
        ],
        "summary": "Mint a key",
        "description": "Requires **admin**. The only response that ever carries the plaintext secret.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The minted key, secret included.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            }
          }
        }
      }
    },
...
    "/llms.txt": {
      "get": {
        "tags": [
          "Redirect"
        ],
        "operationId": "getLlmsTxt",
        "summary": "Published link catalogue",
        "description": "Lists links from the requested host that were explicitly opted in with `listed: true` — archived, expired and unlisted links never appear. No key required: this route sits outside `/api/v1` by design. See `docs/adr/0013`.\n",
        "security": [],
        "responses": {
          "200": {
            "description": "The catalogue, possibly empty.",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The host does not match an active domain."
          }
        }
      },
      "head": {
        "tags": [
          "Redirect"
        ],
        "operationId": "headLlmsTxt",
        "summary": "Published link catalogue, without a body",
        "security": [],
        "responses": {
          "200": {
            "description": "As `GET`, with no body."
          },
          "404": {
            "description": "As `GET`."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer <key>`. The preferred form."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Accepted as a fallback for clients that cannot set `Authorization`."
      }
    },
    "parameters": {
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
...
    },
    "responses": {
...
      "ValidationFailed": {
        "description": "The request failed schema validation. `details` carries the issues.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, unknown or expired key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The caller is known but not allowed to do this.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every error response in linq has this shape. `code` maps one-to-one onto\nthe status: `validation_failed` 400, `unauthorized` 401, `forbidden` 403,\n`not_found` 404, `conflict` 409, `internal` 500.\n",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "validation_failed",
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "conflict",
                  "internal"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "description": "Present on `validation_failed`, carrying the zod issues.",
                "items": {}
              }
            }
          }
        },
        "examples": [
          {
            "error": {
              "code": "not_found",
              "message": "link not found"
            }
          }
        ]
      },
      "Page": {
        "type": "object",
        "description": "The envelope every list shares. `total` counts all matching rows, not the page.",
        "required": [
          "data",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {}
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
...
      "Platform": {
        "type": "string",
        "description": "Derived from the user agent; anything unrecognised is `desktop`.",
        "enum": [
          "android",
          "ios",
          "desktop"
        ]
      },
      "Domain": {
        "type": "object",
        "required": [
          "id",
          "host",
          "fallback_url",
          "base_path_redirect",
          "invalid_short_url_redirect",
          "status",
          "link_count",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "host": {
            "type": "string",
            "maxLength": 253,
            "pattern": "^[a-z0-9.-]+(:\\d{1,5})?$",
            "description": "Lowercased, optionally carrying a port."
          },
          "fallback_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where orphan visits go. Null means 404."
          },
          "base_path_redirect": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where a bare GET / on this host goes. Null falls back to fallback_url."
          },
          "invalid_short_url_redirect": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where a malformed (not just unknown) slug goes. Null falls back to fallback_url."
          },
          "status": {
            "$ref": "#/components/schemas/ResourceStatus"
          },
          "link_count": {
            "type": "integer",
            "description": "Every link on the domain, archived included — the count that must reach zero before the domain may be archived or purged."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DomainCreate": {
        "type": "object",
        "required": [
          "host"
        ],
        "properties": {
          "host": {
            "type": "string",
            "maxLength": 253,
            "pattern": "^[a-z0-9.-]+(:\\d{1,5})?$"
          },
          "fallback_url": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Url"
              },
              {
                "type": "null"
              }
            ],
            "description": "Where an unmatched-but-well-formed slug goes. Absent or null means 404."
          },
          "base_path_redirect": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Url"
              },
              {
                "type": "null"
              }
            ],
            "description": "Where a bare GET / on this host goes. Absent falls back to fallback_url."
          },
          "invalid_short_url_redirect": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Url"
              },
              {
                "type": "null"
              }
            ],
            "description": "Where a malformed (not just unknown) slug goes. Absent falls back to fallback_url."
          }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "description": "An API key: the principal. Revoking one is a real delete. Links carry no reference to the key that created them.\n",
        "required": [
          "id",
          "name",
          "role",
          "prefix",
          "expires_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "Who or what the key is for. Not unique."
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "prefix": {
            "type": "string",
            "description": "The first characters of the secret, so a key is recognisable in a list."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}